requirements.txt
openai>=1.30.0
respan-ai>=0.1.0
respan-instrumentation-openai>=0.1.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
requests>=2.31.0
